refactor: move the tooltip dir one level up
