--- packages/genome-spy/src/fetchers/entrez.js
@@ -1,6 +1,0 @@
-/*
- * https://www.ncbi.nlm.nih.gov/books/NBK25500/
- *
- * TODO: Implement tool & email parameters: https://www.ncbi.nlm.nih.gov/books/NBK25497/
- * TODO: Implement throttling, max three request per second
- */
